Refactor file structures
