--- Monsters.js
@@ -1,5 +1,0 @@
-import Enermy from "./Enermy";
-
-var enermy_1 = new Enermy();
-
-export default { enermy_1 };
