+++ entities/Character.js
@@ -1,0 +1,3 @@
+class Character {
+  constructor() {}
+}
