+++ index.html
@@ -5,8 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Take</title>
-    <link rel="icon" href="O.ico" />
-    <link rel="stylesheet" href="take.css" />
+    <link rel="icon" href="./assets/O.ico" />
+    <link rel="stylesheet" href="./styles/take.css" />
     <script src="take.js" type="module"></script>
   </head>
